% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_to_factor.DS.r
\name{cast_to_factorDS}
\alias{cast_to_factorDS}
\title{Cast specified columns to factor}
\usage{
cast_to_factorDS(df, columns)
}
\arguments{
\item{df}{A data frame containing the columns to be converted.}

\item{columns}{A single character string specifying the column names
separated by "$".}
}
\value{
The input data frame with the specified columns converted to factors.
}
\description{
Convert one or more columns of a data frame to factor type.
}
